Split apart wheel menu code
